+++ packages/nvidia_nat_core/src/nat/front_ends/fastapi/routes/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""FastAPI route registration helpers organized by endpoint area."""
+
