+++ packages/nvidia_nat_core/src/nat/front_ends/fastapi/routes/auth.py
@@ -1,0 +1,67 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""OAuth callback route registration."""
+
+import logging
+from typing import TYPE_CHECKING
+
+import httpx
+from authlib.common.errors import AuthlibBaseError as OAuthError
+from fastapi import FastAPI
+from fastapi import Request
+from fastapi.responses import HTMLResponse
+
+from nat.front_ends.fastapi.html_snippets.auth_code_grant_success import AUTH_REDIRECT_SUCCESS_HTML
+
+if TYPE_CHECKING:
+    from nat.front_ends.fastapi.fastapi_front_end_plugin_worker import FastApiFrontEndPluginWorker
+
+logger = logging.getLogger(__name__)
+
+
+async def add_authorization_route(worker: "FastApiFrontEndPluginWorker", app: FastAPI):
+    """Add OAuth2 callback route for authorization-code flow."""
+
+    async def redirect_uri(request: Request):
+        """Handle the redirect URI for OAuth2 authentication."""
+        state = request.query_params.get("state")
+
+        async with worker._outstanding_flows_lock:
+            if not state or state not in worker._outstanding_flows:
+                return "Invalid state. Please restart the authentication process."
+
+            flow_state = worker._outstanding_flows[state]
+
+        config = flow_state.config
+        verifier = flow_state.verifier
+        client = flow_state.client
+
+        try:
+            res = await client.fetch_token(url=config.token_url,
+                                           authorization_response=str(request.url),
+                                           code_verifier=verifier,
+                                           state=state)
+            flow_state.future.set_result(res)
+        except OAuthError as e:
+            flow_state.future.set_exception(
+                RuntimeError(f"Authorization server rejected request: {e.error} ({e.description})"))
+        except httpx.HTTPError as e:
+            flow_state.future.set_exception(RuntimeError(f"Network error during token fetch: {e}"))
+        except Exception as e:
+            flow_state.future.set_exception(RuntimeError(f"Authentication failed: {e}"))
+
+        return HTMLResponse(content=AUTH_REDIRECT_SUCCESS_HTML,
+                            status_code=200,
+                            headers={
+                                "Content-Type": "text/html; charset=utf-8", "Cache-Control": "no-cache"
+                            })
+
+    if worker.front_end_config.oauth2_callback_path:
+        app.add_api_route(
+            path=worker.front_end_config.oauth2_callback_path,
+            endpoint=redirect_uri,
+            methods=["GET"],
+            description="Handles the authorization code and state returned from the Authorization Code Grant Flow.",
+        )
+
