+++ packages/nvidia_nat_core/tests/nat/front_ends/fastapi/test_job_store.py
@@ -46,7 +46,7 @@ async def failing_job_function() -> None:
 @pytest.mark.asyncio
 async def test_job_store_init_with_engine(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test JobStore initialization with provided database engine."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -57,7 +57,7 @@ async def test_job_store_init_with_engine(db_engine: "AsyncEngine", dask_schedul
 @pytest.mark.asyncio
 async def test_job_store_init_with_db_url(db_url: str, dask_scheduler_address: str):
     """Test JobStore initialization with database URL."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_url=db_url)
 
@@ -67,7 +67,7 @@ async def test_job_store_init_with_db_url(db_url: str, dask_scheduler_address: s
 
 def test_job_store_init_missing_db_params(dask_scheduler_address: str):
     """Test JobStore fails when both db_engine and db_url are missing."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     with pytest.raises(ValueError, match="Either db_engine or db_url must be provided"):
         JobStore(scheduler_address=dask_scheduler_address)
@@ -75,7 +75,7 @@ def test_job_store_init_missing_db_params(dask_scheduler_address: str):
 
 def test_ensure_job_id_with_existing_id(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test ensure_job_id returns the same ID when one is provided."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -86,7 +86,7 @@ def test_ensure_job_id_with_existing_id(db_engine: "AsyncEngine", dask_scheduler
 
 def test_ensure_job_id_generates_new_id(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test ensure_job_id generates a new ID when None is provided."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -100,8 +100,8 @@ def test_ensure_job_id_generates_new_id(db_engine: "AsyncEngine", dask_scheduler
 @pytest.mark.asyncio
 async def test_create_job_default_params(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test job creation with default parameters."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -122,7 +122,7 @@ async def test_create_job_default_params(db_engine: "AsyncEngine", dask_schedule
 @pytest.mark.asyncio
 async def test_create_job_with_params(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test job creation with custom parameters."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -144,7 +144,7 @@ async def test_create_job_with_params(db_engine: "AsyncEngine", dask_scheduler_a
 @pytest.mark.asyncio
 async def test_create_job_clamps_expiry(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test job creation clamps expiry seconds to valid range."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -165,8 +165,8 @@ async def test_create_job_clamps_expiry(db_engine: "AsyncEngine", dask_scheduler
 @pytest.mark.asyncio
 async def test_submit_job_success(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test successful job submission."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -185,7 +185,7 @@ async def test_submit_job_success(db_engine: "AsyncEngine", dask_scheduler_addre
 @pytest.mark.asyncio
 async def test_submit_job_with_sync_timeout(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test job submission with sync timeout to get immediate result."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -204,7 +204,7 @@ async def test_submit_job_with_sync_timeout(db_engine: "AsyncEngine", dask_sched
 @pytest.mark.asyncio
 async def test_submit_job_with_kwargs(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test job submission with keyword arguments."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -225,8 +225,8 @@ async def test_submit_job_with_kwargs(db_engine: "AsyncEngine", dask_scheduler_a
 @pytest.mark.asyncio
 async def test_update_status_basic(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test basic status update."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -251,8 +251,8 @@ async def test_update_status_basic(db_engine: "AsyncEngine", dask_scheduler_addr
 @pytest.mark.asyncio
 async def test_update_status_with_error(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test status update with error message."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -271,8 +271,8 @@ async def test_update_status_with_error(db_engine: "AsyncEngine", dask_scheduler
 @pytest.mark.asyncio
 async def test_update_status_with_pydantic_output(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test status update with Pydantic model output."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -295,8 +295,8 @@ async def test_update_status_with_pydantic_output(db_engine: "AsyncEngine", dask
 @pytest.mark.asyncio
 async def test_update_status_with_dict_output(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test status update with dictionary output."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -319,8 +319,8 @@ async def test_update_status_with_dict_output(db_engine: "AsyncEngine", dask_sch
 @pytest.mark.asyncio
 async def test_update_status_nonexistent_job(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test updating status of non-existent job raises error."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -332,8 +332,8 @@ async def test_update_status_nonexistent_job(db_engine: "AsyncEngine", dask_sche
 @pytest.mark.asyncio
 async def test_get_job_existing(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting an existing job."""
-    from nat.front_ends.fastapi.job_store import JobInfo
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobInfo
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -349,7 +349,7 @@ async def test_get_job_existing(db_engine: "AsyncEngine", dask_scheduler_address
 @pytest.mark.asyncio
 async def test_get_job_nonexistent(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting a non-existent job returns None."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -361,8 +361,8 @@ async def test_get_job_nonexistent(db_engine: "AsyncEngine", dask_scheduler_addr
 @pytest.mark.asyncio
 async def test_get_status_existing(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting status of an existing job."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -376,8 +376,8 @@ async def test_get_status_existing(db_engine: "AsyncEngine", dask_scheduler_addr
 @pytest.mark.asyncio
 async def test_get_status_nonexistent(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting status of non-existent job returns NOT_FOUND."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -389,7 +389,7 @@ async def test_get_status_nonexistent(db_engine: "AsyncEngine", dask_scheduler_a
 @pytest.mark.asyncio
 async def test_get_all_jobs_empty(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting all jobs when database is empty."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -401,7 +401,7 @@ async def test_get_all_jobs_empty(db_engine: "AsyncEngine", dask_scheduler_addre
 @pytest.mark.asyncio
 async def test_get_all_jobs_multiple(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting all jobs with multiple jobs in database."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -421,7 +421,7 @@ async def test_get_all_jobs_multiple(db_engine: "AsyncEngine", dask_scheduler_ad
 @pytest.mark.asyncio
 async def test_get_last_job_empty(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting last job when database is empty."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -433,7 +433,7 @@ async def test_get_last_job_empty(db_engine: "AsyncEngine", dask_scheduler_addre
 @pytest.mark.asyncio
 async def test_get_last_job_multiple(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test getting last job with multiple jobs in database."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -453,8 +453,8 @@ async def test_get_last_job_multiple(db_engine: "AsyncEngine", dask_scheduler_ad
 @pytest.mark.asyncio
 async def test_get_jobs_by_status(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test filtering jobs by status."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -488,8 +488,8 @@ async def test_get_jobs_by_status(db_engine: "AsyncEngine", dask_scheduler_addre
 @pytest.mark.asyncio
 async def test_get_expires_at_active_job(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test get_expires_at for active jobs returns None."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -514,8 +514,8 @@ async def test_get_expires_at_active_job(db_engine: "AsyncEngine", dask_schedule
 @pytest.mark.asyncio
 async def test_get_expires_at_finished_job(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test get_expires_at for finished jobs returns correct expiry time."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -543,8 +543,8 @@ async def test_get_expires_at_finished_job(db_engine: "AsyncEngine", dask_schedu
 @pytest.mark.asyncio
 async def test_cleanup_expired_jobs_no_expired(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test cleanup when no jobs are expired."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -575,8 +575,8 @@ async def test_cleanup_expired_jobs_with_output_files(db_engine: "AsyncEngine",
                                                       tmp_path: Path,
                                                       monkeypatch: pytest.MonkeyPatch):
     """Test cleanup removes output files for expired jobs."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     with monkeypatch.context() as monkey_context:
         # Lower minimum expiry for testing
@@ -628,8 +628,8 @@ async def test_cleanup_expired_jobs_keeps_active(db_engine: "AsyncEngine",
                                                  dask_scheduler_address: str,
                                                  monkeypatch: pytest.MonkeyPatch):
     """Test cleanup never expires active (running/submitted) jobs."""
-    from nat.front_ends.fastapi.job_store import JobStatus
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStatus
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     with monkeypatch.context() as monkey_context:
         # Lower minimum expiry for testing
@@ -667,7 +667,7 @@ async def test_cleanup_expired_jobs_keeps_active(db_engine: "AsyncEngine",
 
 def test_get_db_engine_with_url():
     """Test get_db_engine with provided URL."""
-    from nat.front_ends.fastapi.job_store import get_db_engine
+    from nat.front_ends.fastapi.async_jobs import get_db_engine
 
     db_url = "sqlite:///test.db"
     engine = get_db_engine(db_url, use_async=False)
@@ -678,7 +678,7 @@ def test_get_db_engine_with_url():
 
 def test_get_db_engine_async():
     """Test get_db_engine creates async engine."""
-    from nat.front_ends.fastapi.job_store import get_db_engine
+    from nat.front_ends.fastapi.async_jobs import get_db_engine
 
     db_url = "sqlite+aiosqlite:///test.db"
     engine = get_db_engine(db_url, use_async=True)
@@ -690,7 +690,7 @@ def test_get_db_engine_async():
 
 def test_get_db_engine_from_env_var(set_nat_job_store_db_url_env_var: str):
     """Test get_db_engine uses environment variable when no URL provided."""
-    from nat.front_ends.fastapi.job_store import get_db_engine
+    from nat.front_ends.fastapi.async_jobs import get_db_engine
 
     engine = get_db_engine(use_async=True)
 
@@ -701,7 +701,7 @@ def test_get_db_engine_from_env_var(set_nat_job_store_db_url_env_var: str):
 
 def test_get_db_engine_creates_default_sqlite():
     """Test get_db_engine creates default SQLite when no URL provided."""
-    from nat.front_ends.fastapi.job_store import get_db_engine
+    from nat.front_ends.fastapi.async_jobs import get_db_engine
 
     # Temporarily clear the environment variable
     original_url = os.environ.get("NAT_JOB_STORE_DB_URL")
@@ -726,7 +726,7 @@ def test_job_store_dask_client_property(dask_client: "DaskClient",
                                         db_engine: "AsyncEngine",
                                         dask_scheduler_address: str):
     """Test the dask_client property works correctly."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
 
@@ -738,7 +738,7 @@ def test_job_store_dask_client_property(dask_client: "DaskClient",
 @pytest.mark.asyncio
 async def test_session_context_manager(db_engine: "AsyncEngine", dask_scheduler_address: str):
     """Test the session context manager works correctly."""
-    from nat.front_ends.fastapi.job_store import JobStore
+    from nat.front_ends.fastapi.async_jobs import JobStore
 
     job_store = JobStore(scheduler_address=dask_scheduler_address, db_engine=db_engine)
